backend/requirements.txt
# FastAPI Backend Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# WebSocket Support
websockets==12.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# File Upload & Storage
aiofiles==23.2.1
python-magic==0.4.27
Pillow==10.1.0

# Utils
httpx==0.25.1
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

